Local and MythTV bookmark support
A preference setting allows selection of local, mythtv, or auto selection.
Local will store the bookmarks on a local database. This works always.
MythTV will store the bookmark as a MythTV bookmark on the server. This
only works where there is a seektable present.
Auto selection will store bookmark on MythTV and if it cannot, it will use the
local database.

+++ app/src/main/java/org/mythtv/leanfront/data/VideoDbHelper.java
@@ -17,13 +17,12 @@
 package org.mythtv.leanfront.data;
 
 import android.content.Context;
-import android.content.SharedPreferences;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
-import androidx.preference.PreferenceManager;
-
 import org.mythtv.leanfront.data.VideoContract.VideoEntry;
+import org.mythtv.leanfront.data.VideoContract.StatusEntry;
+
 
 /**
  * VideoDbHelper manages the creation and upgrade of the database used in this sample.
@@ -31,7 +30,7 @@ import org.mythtv.leanfront.data.VideoContract.VideoEntry;
 public class VideoDbHelper extends SQLiteOpenHelper {
 
     // Change this when you change the database schema.
-    private static final int DATABASE_VERSION = 7;
+    private static final int DATABASE_VERSION = 1;
 
     // The name of our database.
     private static final String DATABASE_NAME = "leanback.db";
@@ -42,43 +41,55 @@ public class VideoDbHelper extends SQLiteOpenHelper {
 
     @Override
     public void onCreate(SQLiteDatabase db) {
-        // Create a table to hold videos.
-        final String SQL_CREATE_VIDEO_TABLE = "CREATE TABLE " + VideoEntry.TABLE_NAME + " (" +
-                VideoEntry._ID + " INTEGER PRIMARY KEY," +
-                VideoEntry.COLUMN_TITLE + " TEXT NOT NULL, " +
-                VideoEntry.COLUMN_SUBTITLE + " TEXT NOT NULL, " +
-                VideoEntry.COLUMN_VIDEO_URL + " TEXT NOT NULL, " +
-                VideoEntry.COLUMN_DESC + " TEXT NOT NULL, " +
-                VideoEntry.COLUMN_BG_IMAGE_URL + " TEXT NOT NULL, " +
-                VideoEntry.COLUMN_STUDIO + " TEXT NOT NULL, " +
-                VideoEntry.COLUMN_CARD_IMG + " TEXT NOT NULL, " +
-                VideoEntry.COLUMN_CONTENT_TYPE + " TEXT NOT NULL, " +
-                VideoEntry.COLUMN_PRODUCTION_YEAR + " TEXT, " +
-                VideoEntry.COLUMN_DURATION + " TEXT NOT NULL, " +
-                VideoEntry.COLUMN_ACTION + " TEXT NOT NULL," +
-                VideoEntry.COLUMN_AIRDATE + " TEXT," +
-                VideoEntry.COLUMN_STARTTIME + " TEXT," +
-                VideoEntry.COLUMN_RECORDEDID + " TEXT," +
-                VideoEntry.COLUMN_STORAGEGROUP + " TEXT," +
-                VideoEntry.COLUMN_RECGROUP + " TEXT," +
-                VideoEntry.COLUMN_SEASON + " TEXT," +
-                VideoEntry.COLUMN_EPISODE + " TEXT" +
-                " );";
-
-        // Do the creating of the databases.
-        db.execSQL(SQL_CREATE_VIDEO_TABLE);
+        onUpgrade(db,0,DATABASE_VERSION);
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        // Simply discard all old data and start over when upgrading.
-        db.execSQL("DROP TABLE IF EXISTS " + VideoEntry.TABLE_NAME);
-        onCreate(db);
+        if (oldVersion == 0) {
+            db.execSQL("DROP TABLE IF EXISTS " + VideoEntry.TABLE_NAME);
+            // Create a table to hold videos.
+            // This table gets deleted and recreated periodically
+            final String SQL_CREATE_VIDEO_TABLE = "CREATE TABLE " + VideoEntry.TABLE_NAME + " (" +
+                    VideoEntry._ID + " INTEGER PRIMARY KEY," +
+                    VideoEntry.COLUMN_TITLE + " TEXT NOT NULL, " +
+                    VideoEntry.COLUMN_SUBTITLE + " TEXT NOT NULL, " +
+                    VideoEntry.COLUMN_VIDEO_URL + " TEXT NOT NULL, " +
+                    VideoEntry.COLUMN_DESC + " TEXT NOT NULL, " +
+                    VideoEntry.COLUMN_BG_IMAGE_URL + " TEXT NOT NULL, " +
+                    VideoEntry.COLUMN_STUDIO + " TEXT NOT NULL, " +
+                    VideoEntry.COLUMN_CARD_IMG + " TEXT NOT NULL, " +
+                    VideoEntry.COLUMN_CONTENT_TYPE + " TEXT NOT NULL, " +
+                    VideoEntry.COLUMN_PRODUCTION_YEAR + " TEXT, " +
+                    VideoEntry.COLUMN_DURATION + " TEXT NOT NULL, " +
+                    VideoEntry.COLUMN_ACTION + " TEXT NOT NULL," +
+                    VideoEntry.COLUMN_AIRDATE + " TEXT," +
+                    VideoEntry.COLUMN_STARTTIME + " TEXT," +
+                    VideoEntry.COLUMN_RECORDEDID + " TEXT," +
+                    VideoEntry.COLUMN_STORAGEGROUP + " TEXT," +
+                    VideoEntry.COLUMN_RECGROUP + " TEXT," +
+                    VideoEntry.COLUMN_SEASON + " TEXT," +
+                    VideoEntry.COLUMN_EPISODE + " TEXT" +
+                    " );";
+
+            // Do the creating of the table.
+            db.execSQL(SQL_CREATE_VIDEO_TABLE);
+
+            db.execSQL("DROP TABLE IF EXISTS " + StatusEntry.TABLE_NAME);
+            // videostatus table keeps bookmarks even when video table is reloaded.
+            // LAST_USED column is datetime, used to delete entries older than a month.
+            final String SQL_CREATE_VIDEOSTATUS_TABLE = "CREATE TABLE " + StatusEntry.TABLE_NAME + " (" +
+                    StatusEntry._ID + " INTEGER PRIMARY KEY," +
+                    StatusEntry.COLUMN_VIDEO_URL + " TEXT NOT NULL UNIQUE, " +
+                    StatusEntry.COLUMN_LAST_USED + " INTEGER NOT NULL, " +
+                    StatusEntry.COLUMN_BOOKMARK + " INTEGER);";
+            db.execSQL(SQL_CREATE_VIDEOSTATUS_TABLE);
+            oldVersion = 1;
+        }
     }
 
     @Override
     public void onDowngrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        // Do the same thing as upgrading...
-        onUpgrade(db, oldVersion, newVersion);
+        onCreate(db);
     }
 }
